Adding iOS support to Swift Package Manager builds. Had to restructure the Apple-specific files to do this.
