+++ Package.swift
@@ -2,17 +2,9 @@
 
 import PackageDescription
 
-#if os(macOS)
-let platformDepedencies: [Package.Dependency] = []
-let platformExcludes = ["iOS", "Linux", "Operations/Shaders"]
-let platformTargets: [Target] = [
-        .target(
-            name: "GPUImage",
-            path: "framework/Source",
-            exclude: platformExcludes)]
-#elseif os(iOS)
-let platformDepedencies: [Package.Dependency] = []
-let platformExcludes = ["Linux", "Mac", "Operations/Shaders"]
+#if os(macOS) // This fires for both macOS and iOS targets, because it's based on build platform
+let platformDependencies: [Package.Dependency] = []
+let platformExcludes = ["Linux", "Operations/Shaders"]
 let platformTargets: [Target] = [
         .target(
             name: "GPUImage",
@@ -21,11 +13,11 @@ let platformTargets: [Target] = [
 #elseif os(Linux)
 // TODO: Add back in RPi support
 // TODO: Move the remote system library packages into this project
-let platformDepedencies: [Package.Dependency] = [
+let platformDependencies: [Package.Dependency] = [
     .package(url: "https://github.com/BradLarson/COpenGL.git", from: "1.0.2"), 
     .package(url: "https://github.com/BradLarson/CFreeGLUT.git", from: "1.0.1"), 
     .package(url: "https://github.com/BradLarson/CVideo4Linux.git", from: "1.0.2")]
-let platformExcludes =  ["iOS", "Mac", "Operations/Shaders", "Linux/RPiRenderWindow.swift", "Linux/OpenGLContext-RPi.swift", "Linux/V4LSupplement"]
+let platformExcludes =  ["Apple", "Operations/Shaders", "Linux/RPiRenderWindow.swift", "Linux/OpenGLContext-RPi.swift", "Linux/V4LSupplement"]
 let platformTargets: [Target] = [
         .target(
             name: "V4LSupplement",
@@ -45,7 +37,7 @@ let package = Package(
             name: "GPUImage",
             targets: ["GPUImage"]),
     ],
-    dependencies: platformDepedencies,
+    dependencies: platformDependencies,
     targets: platformTargets,
     swiftLanguageVersions: [.v4]
 )
