+++ framework/Source/Apple/PictureOutput.swift
@@ -1,5 +1,16 @@
+#if canImport(OpenGL)
 import OpenGL.GL3
+public typealias PlatformImageType = NSImage
+#else
+import OpenGLES
+public typealias PlatformImageType = UIImage
+#endif
+
+#if canImport(UIKit)
+import UIKit
+#else
 import Cocoa
+#endif
 
 public enum PictureFileFormat {
     case png
@@ -9,8 +20,10 @@ public enum PictureFileFormat {
 public class PictureOutput: ImageConsumer {
     public var encodedImageAvailableCallback:((Data) -> ())?
     public var encodedImageFormat:PictureFileFormat = .png
-    public var imageAvailableCallback:((NSImage) -> ())?
+    public var imageAvailableCallback:((PlatformImageType) -> ())?
     public var onlyCaptureNextFrame:Bool = true
+    public var keepImageAroundForSynchronousCapture:Bool = false
+    var storedFramebuffer:Framebuffer?
     
     public let sources = SourceContainer()
     public let maximumInputs:UInt = 1
@@ -21,53 +34,54 @@ public class PictureOutput: ImageConsumer {
     
     deinit {
     }
-
+    
     public func saveNextFrameToURL(_ url:URL, format:PictureFileFormat) {
         onlyCaptureNextFrame = true
         encodedImageFormat = format
         self.url = url // Create an intentional short-term retain cycle to prevent deallocation before next frame is captured
         encodedImageAvailableCallback = {imageData in
             do {
-// FIXME: Xcode 8 beta 2
                 try imageData.write(to: self.url, options:.atomic)
-//                try imageData.write(to: self.url, options:NSData.WritingOptions.dataWritingAtomic)
             } catch {
                 // TODO: Handle this better
                 print("WARNING: Couldn't save image with error:\(error)")
             }
         }
     }
-
-    // TODO: Replace with texture caches and a safer capture routine
+    
+    // TODO: Replace with texture caches
     func cgImageFromFramebuffer(_ framebuffer:Framebuffer) -> CGImage {
         let renderFramebuffer = sharedImageProcessingContext.framebufferCache.requestFramebufferWithProperties(orientation:framebuffer.orientation, size:framebuffer.size)
         renderFramebuffer.lock()
         renderFramebuffer.activateFramebufferForRendering()
-        clearFramebufferWithColor(Color.transparent)
-
-        // Need the blending here to enable non-1.0 alpha on output image
-        enableAdditiveBlending()
-        
+        clearFramebufferWithColor(Color.red)
         renderQuadWithShader(sharedImageProcessingContext.passthroughShader, uniformSettings:ShaderUniformSettings(), vertexBufferObject:sharedImageProcessingContext.standardImageVBO, inputTextures:[framebuffer.texturePropertiesForOutputRotation(.noRotation)])
-
-        disableBlending()
-        
         framebuffer.unlock()
         
         let imageByteSize = Int(framebuffer.size.width * framebuffer.size.height * 4)
-        let data = UnsafeMutablePointer<UInt8>.allocate(capacity:imageByteSize)
+        let data = UnsafeMutablePointer<UInt8>.allocate(capacity: imageByteSize)
         glReadPixels(0, 0, framebuffer.size.width, framebuffer.size.height, GLenum(GL_RGBA), GLenum(GL_UNSIGNED_BYTE), data)
         renderFramebuffer.unlock()
-        guard let dataProvider = CGDataProvider(dataInfo: nil, data: data, size: imageByteSize, releaseData: dataProviderReleaseCallback) else {fatalError("Could not create CGDataProvider")}
+        guard let dataProvider = CGDataProvider(dataInfo:nil, data:data, size:imageByteSize, releaseData: dataProviderReleaseCallback) else {fatalError("Could not allocate a CGDataProvider")}
         let defaultRGBColorSpace = CGColorSpaceCreateDeviceRGB()
-        
-      return CGImage(width: Int(framebuffer.size.width), height: Int(framebuffer.size.height), bitsPerComponent:8, bitsPerPixel:32, bytesPerRow:4 * Int(framebuffer.size.width), space:defaultRGBColorSpace, bitmapInfo:CGBitmapInfo(rawValue: CGImageAlphaInfo.premultipliedLast.rawValue), provider:dataProvider, decode:nil, shouldInterpolate:false, intent:.defaultIntent)!
+        return CGImage(width:Int(framebuffer.size.width), height:Int(framebuffer.size.height), bitsPerComponent:8, bitsPerPixel:32, bytesPerRow:4 * Int(framebuffer.size.width), space:defaultRGBColorSpace, bitmapInfo:CGBitmapInfo() /*| CGImageAlphaInfo.Last*/, provider:dataProvider, decode:nil, shouldInterpolate:false, intent:.defaultIntent)!
     }
     
     public func newFramebufferAvailable(_ framebuffer:Framebuffer, fromSourceIndex:UInt) {
+        if keepImageAroundForSynchronousCapture {
+            storedFramebuffer?.unlock()
+            storedFramebuffer = framebuffer
+        }
+        
         if let imageCallback = imageAvailableCallback {
             let cgImageFromBytes = cgImageFromFramebuffer(framebuffer)
+            
+            // TODO: Let people specify orientations
+#if canImport(UIKit)
+            let image = UIImage(cgImage:cgImageFromBytes, scale:1.0, orientation:.up)
+#else
             let image = NSImage(cgImage:cgImageFromBytes, size:NSZeroSize)
+#endif
             
             imageCallback(image)
             
@@ -78,12 +92,21 @@ public class PictureOutput: ImageConsumer {
         
         if let imageCallback = encodedImageAvailableCallback {
             let cgImageFromBytes = cgImageFromFramebuffer(framebuffer)
-            let bitmapRepresentation = NSBitmapImageRep(cgImage:cgImageFromBytes)
             let imageData:Data
+            
+#if canImport(UIKit)
+            let image = UIImage(cgImage:cgImageFromBytes, scale:1.0, orientation:.up)
+            switch encodedImageFormat {
+                case .png: imageData = UIImagePNGRepresentation(image)! // TODO: Better error handling here
+                case .jpeg: imageData = UIImageJPEGRepresentation(image, 0.8)! // TODO: Be able to set image quality
+            }
+#else
+            let bitmapRepresentation = NSBitmapImageRep(cgImage:cgImageFromBytes)
             switch encodedImageFormat {
                 case .png: imageData = bitmapRepresentation.representation(using: .png, properties: [NSBitmapImageRep.PropertyKey(rawValue: ""):""])!
                 case .jpeg: imageData = bitmapRepresentation.representation(using: .jpeg, properties: [NSBitmapImageRep.PropertyKey(rawValue: ""):""])!
             }
+#endif
 
             imageCallback(imageData)
             
@@ -92,6 +115,20 @@ public class PictureOutput: ImageConsumer {
             }
         }
     }
+    
+#if canImport(UIKit)
+    public func synchronousImageCapture() -> UIImage {
+        var outputImage:UIImage!
+        sharedImageProcessingContext.runOperationSynchronously{
+            guard let currentFramebuffer = storedFramebuffer else { fatalError("Synchronous access requires keepImageAroundForSynchronousCapture to be set to true") }
+            
+            let cgImageFromBytes = cgImageFromFramebuffer(currentFramebuffer)
+            outputImage = UIImage(cgImage:cgImageFromBytes, scale:1.0, orientation:.up)
+        }
+        
+        return outputImage
+    }
+#endif
 }
 
 public extension ImageSource {
@@ -102,16 +139,16 @@ public extension ImageSource {
     }
 }
 
-public extension NSImage {
-    func filterWithOperation<T:ImageProcessingOperation>(_ operation:T) -> NSImage {
+public extension PlatformImageType {
+    func filterWithOperation<T:ImageProcessingOperation>(_ operation:T) -> PlatformImageType {
         return filterWithPipeline{input, output in
             input --> operation --> output
         }
     }
-
-    func filterWithPipeline(_ pipeline:(PictureInput, PictureOutput) -> ()) -> NSImage {
+    
+    func filterWithPipeline(_ pipeline:(PictureInput, PictureOutput) -> ()) -> PlatformImageType {
+        var outputImage:PlatformImageType?
         let picture = PictureInput(image:self)
-        var outputImage:NSImage?
         let pictureOutput = PictureOutput()
         pictureOutput.onlyCaptureNextFrame = true
         pictureOutput.imageAvailableCallback = {image in
@@ -125,7 +162,5 @@ public extension NSImage {
 
 // Why are these flipped in the callback definition?
 func dataProviderReleaseCallback(_ context:UnsafeMutableRawPointer?, data:UnsafeRawPointer, size:Int) {
-//    UnsafeMutablePointer<UInt8>(data).deallocate(capacity:size)
-    // FIXME: Verify this is correct
     data.deallocate()
 }
